+++ opendut-model/src/peer/mod.rs
@@ -318,7 +318,7 @@ mod tests {
             },
             vpn: VpnPeerConfiguration::Netbird {
                 management_url: Url::parse("https://netbird.opendut.local/api")?,
-                setup_key: SetupKey::from(Uuid::parse_str("d79c202f-bbbf-4997-844e-678f27606e1c")?),
+                setup_key: SetupKey::new("d79c202f-bbbf-4997-844e-678f27606e1c"),
             },
         };